--- server/src/main/java/com/sju18/PetManagement/domain/account/dto/LoginRequestDTO.java
@@ -1,11 +1,0 @@
-package com.sju18.PetManagement.domain.account.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginRequestDTO {
-
-    private String username;
-    private String password;
-
-}
