--- server/src/main/java/com/sju18/PetManagement/domain/account/dto/LoginResponseDTO.java
@@ -1,12 +1,0 @@
-package com.sju18.PetManagement.domain.account.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class LoginResponseDTO {
-
-    private String token;
-
-}
