--- server/src/main/java/com/sju18/PetManagement/domain/account/dto/ProfileUpdateResponseDTO.java
@@ -1,10 +1,0 @@
-package com.sju18.PetManagement.domain.account.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ProfileUpdateResponseDTO {
-    String message;
-}
