--- server/src/main/java/com/sju18/PetManagement/domain/account/dto/ResignRequestDTO.java
@@ -1,7 +1,0 @@
-package com.sju18.PetManagement.domain.account.dto;
-
-import lombok.Data;
-
-@Data
-public class ResignRequestDTO {
-}
