--- server/src/main/java/com/sju18/PetManagement/domain/account/dto/SignupResponseDTO.java
@@ -1,10 +1,0 @@
-package com.sju18.PetManagement.domain.account.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class SignupResponseDTO {
-    private String message;
-}
