--- server/src/main/java/com/sju18/PetManagement/global/config/security/JwtAuthenticationProvider.java
@@ -1,43 +1,0 @@
-package com.sju18.PetManagement.global.config.security;
-
-import com.sju18.PetManagement.domain.account.dao.Account;
-import com.sju18.PetManagement.domain.account.dao.AccountRepository;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class JwtAuthenticationProvider implements AuthenticationProvider {
-
-    @Autowired
-    private AccountRepository accountRepository;
-
-    @Autowired
-    private PasswordEncoder passwordEncoder;
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        String username = authentication.getName();
-        String password = authentication.getCredentials().toString();
-
-        Account account = accountRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException(username));
-
-        if(passwordEncoder.matches(account.getPassword(), password)) {
-            throw new BadCredentialsException("UnAuthorized");
-        }
-
-        return new UsernamePasswordAuthenticationToken(username, password);
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return authentication.equals(UsernamePasswordAuthenticationToken.class);
-    }
-}
