--- server/src/main/java/com/sju18/PetManagement/global/config/security/JwtRequestFilter.java
@@ -1,84 +1,0 @@
-package com.sju18.PetManagement.global.config.security;
-
-import com.sju18.PetManagement.domain.account.application.JwtUserDetailsService;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-@Component
-public class JwtRequestFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private JwtUserDetailsService jwtUserDetailService;
-
-    @Autowired
-    private JwtTokenUtil jwtTokenUtil;
-
-    // 비회원 유저도 접속 허용하는 URL List
-    private static final List<String> EXCLUDE_URL =
-            Collections.unmodifiableList(
-                    Arrays.asList(
-                            "/api/account/signup",
-                            "/api/account/login"
-                    ));
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request,
-                                    HttpServletResponse response,
-                                    FilterChain filterChain) throws ServletException, IOException {
-
-        // request 헤더에서 Authorization 필드를 검출
-        final String requestTokenHeader = request.getHeader("Authorization");
-
-        String username = null;
-        String jwtToken = null;
-
-        if (requestTokenHeader != null) {
-            jwtToken = requestTokenHeader.substring(7);
-            try {
-                username = jwtTokenUtil.getUsernameFromToken(jwtToken);
-            } catch (Exception e) {
-                logger.warn(e.getMessage());
-            }
-        } else {
-            logger.warn("Authorization Header is null");
-        }
-
-        if(username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-            try {
-                UserDetails userDetails = this.jwtUserDetailService.loadUserByUsername(username);
-                if(jwtTokenUtil.validateToken(jwtToken, userDetails)) {
-                    UsernamePasswordAuthenticationToken authenticationToken =
-                            new UsernamePasswordAuthenticationToken(userDetails, null ,userDetails.getAuthorities());
-
-                    authenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                    SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-                }
-            } catch (UsernameNotFoundException e) {
-                System.out.println("User Not Found");
-            }
-        }
-        filterChain.doFilter(request,response);
-    }
-
-    @Override
-    protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
-        return EXCLUDE_URL.stream().anyMatch(exclude -> exclude.equalsIgnoreCase(request.getServletPath()));
-    }
-
-}
