--- server/src/main/java/com/sju18/PetManagement/global/config/security/JwtTokenUtil.java
@@ -1,74 +1,0 @@
-package com.sju18.PetManagement.global.config.security;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Encoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-@Component
-public class JwtTokenUtil {
-
-    // 토큰 생성용 난수 시드값
-    private static final String secret = "newcentury1234567890";
-    private SecretKey secretKey = Keys.secretKeyFor(SignatureAlgorithm.HS512);
-    String base64SecretKey = Encoders.BASE64.encode(secretKey.getEncoded());
-
-    // 인증 토큰 유효기간 (6시간)
-    public static final long JWT_TOKEN_VALIDITY = 6 * 60 * 60;
-
-    public String getUsernameFromToken(String token) {
-        return getClaimFromToken(token, Claims::getId);
-    }
-
-    public <T> T getClaimFromToken(String token, Function<Claims, T> claimsResolver) {
-        final Claims claims = getAllClaimsFromToken(token);
-        return claimsResolver.apply(claims);
-    }
-
-    private Claims getAllClaimsFromToken(String token) {
-        return Jwts.parserBuilder().setSigningKey(secretKey).build().parseClaimsJws(token).getBody();
-    }
-
-    private Boolean isTokenExpired(String token) {
-        final Date expiration = getExpirationDateFromToken(token);
-        return expiration.before(new Date());
-    }
-
-    public Date getExpirationDateFromToken(String token) {
-        return getClaimFromToken(token, Claims::getExpiration);
-    }
-
-    public String generateToken(String id) {
-        return generateToken(id, new HashMap<>());
-    }
-
-    public String generateToken(String id, Map<String, Object> claims) {
-        return doGenerateToken(id, claims);
-    }
-
-    private String doGenerateToken(String id, Map<String, Object> claims) {
-        return Jwts.builder()
-                .setClaims(claims)
-                .setId(id)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + JWT_TOKEN_VALIDITY * 1000))
-                .signWith(secretKey, SignatureAlgorithm.HS512)
-                .compact();
-    }
-
-    public Boolean validateToken(String token, UserDetails userDetails) {
-        final String username = getUsernameFromToken(token);
-        return (username.equals(userDetails.getUsername())) && !isTokenExpired(token);
-    }
-
-}
