+++ server/src/main/java/com/sju18/petmanagement/domain/account/dao/Permission.java
@@ -1,0 +1,15 @@
+package com.sju18.petmanagement.domain.account.dao;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Permission {
+
+    USER("ROLE_USER"),
+    ADMIN("ROLE_ADMIN");
+
+    private final String value;
+
+}
