+++ server/src/main/java/com/sju18/petmanagement/domain/account/dto/LoginRequestDTO.java
@@ -1,0 +1,11 @@
+package com.sju18.petmanagement.domain.account.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginRequestDTO {
+
+    private String username;
+    private String password;
+
+}
