+++ server/src/main/java/com/sju18/petmanagement/domain/account/dto/LoginResponseDTO.java
@@ -1,0 +1,12 @@
+package com.sju18.petmanagement.domain.account.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class LoginResponseDTO {
+
+    private String token;
+
+}
