+++ server/src/main/java/com/sju18/petmanagement/domain/account/dto/ProfileLookupRequestDTO.java
@@ -1,0 +1,7 @@
+package com.sju18.petmanagement.domain.account.dto;
+
+import lombok.Data;
+
+@Data
+public class ProfileLookupRequestDTO {
+}
