+++ server/src/main/java/com/sju18/petmanagement/domain/account/dto/SignupResponseDTO.java
@@ -1,0 +1,10 @@
+package com.sju18.petmanagement.domain.account.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class SignupResponseDTO {
+    private String message;
+}
