--- server/src/test/java/com/sju18/PetManagement/PetManagementApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sju18.PetManagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PetManagementApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
