+++ server/src/test/java/com/sju18/petmanagement/PetManagementApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sju18.petmanagement;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PetManagementApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
